文章新增界面优化

+++ codelife-service/src/main/java/com/liuqi/commons/db/BaseRepository.java
@@ -12,6 +12,6 @@ import org.springframework.data.repository.NoRepositoryBean;
  * @version V1.0
  **/
 @NoRepositoryBean
-public interface BaseRepository<E> extends JpaRepository<E, Long>, JpaSpecificationExecutor<E> {
+public interface BaseRepository<E> extends JpaRepository<E, Integer>, JpaSpecificationExecutor<E> {
 }
